refactor: standardize equipment bulk actions constant filename casing
