+++ src/app/(app)/equipment/_constants/equipmentBulkActions.ts
